Added memory manager
Added a custom memory manager that wraps malloc and, when a debug build,
records the file and line number of an allocation to aid in discovering
memory leaks. Also reverted the type typedefs to long long and not
int64_t so casts are not needed when using library functions or dealing
with literal strings. Removed dependency on SAL.