Tweaked LinkedList
Added a default iterator to the list so that one does not need to be
created when iterating over the entire list in a thread safe manner.
Also added forward declarations for LinkedList structs and tweaked the
iteration macros to take the default iterator into account. Also changed
Insert to operate on an iterator.